fix typing extractor not handling IntEnum handler
Addresses GitHub #9.

Adding a custom handler for IntEnum in TypingExtractor
resulted in the int handler still being used, due to the
handler resolver logic iterating through a list and checking
for matches via seeing if the passed type is a subclass, rather than
looking for the most granular match to the type it can.

Refactoring the logic to iterate the mro and find a match, which
means it'll check for the most granular match first.