removing py2.7 ci, updating black
Black was updated due to a bug that caused it to fail when
using click.

Python2.7 is being removed because the attrs tests now use
proper type annotations. Python2 was having trouble parsing that
syntax, and since the community is moving to Python3, it is
best to stop supporting it.